refactor(c): /simplify — four review passes, applied
Reuse, simplification, efficiency and altitude, reviewed independently and
deduped. What was fixed:

- **`raised` was `Effect.fromResult`.** Three of the four reviews caught it
  independently. A hand-rolled copy of a library primitive, with a docstring
  insisting it was "not a general Result-to-Effect helper" while being exactly
  that. The prose that was worth keeping is now a comment at the one call site;
  the `Result` import it existed to serve is gone.
- **`bytesOf` was fifteen lines of UTF-8 arithmetic** — a surrogate-pair branch
  and a lone-surrogate rule, untested, on the floor, to avoid one transient
  buffer on an occasional agent call. It is `TextEncoder` now. Subtle arithmetic
  that is wrong for years is the more expensive thing. Its cost — a listing is
  O(the bytes of every served `.md`), a class up from the outline listing's
  O(nodes) — is now STATED, with the cheaper form named (`@olai/store` already
  has the byte count and discards it) rather than left for a reader to discover.
- **`noSuchDocument` came back off the floor.** It was the only place in
  `@olai/format` that CONSTRUCTED a failure; the format declares what a refusal
  is and `@olai/ops` is the only layer that raises one. It now sits beside
  `notFound` and `notANode`, which are that layer's other exported refusals, and
  is exported the same way. Facts down, voice up.
- **The bulk broken index has a name.** `new Map(set.broken.map(…))` was written
  twice in one file, identically, and the second was mine — `brokenBy`, beside
  the per-path `brokenIn` the last pass named.
- **The `asking` docstring still said "the four envelopes"** over six of them.
- A dead cucumber step (defined, referenced by no scenario), a second
  `withTools` boot to assert two annotations the existing annotation test was
  already shaped for, a comment pointing at the file this branch deletes, and
  two needless backtick escapes in double-quoted strings.

**Said out loud rather than changed: an agent now has two routes to a body** —
`read_document` and the `documents` collection resource — and they answer
differently on purpose (a `.html` key is reachable on the collection, a broken
file reads there as the empty text the set carries, and there is no near-miss
refusal). That is render-shaped versus request-shaped, which `surface/ops.ts`
argued for the LISTING and not for the BODY. It argues both now. Collapsing
them would take a published URI away from agents or push a refusal into the
member the browser draws from, and neither is this PR's call to make.

Skipped, with reasons: carrying the byte count from `@olai/store` (the store
and the wire `Document` shape, well outside this diff — named in the docstring
instead); `@olai/chat`'s `weight` and `@olai/web`'s `plainLine`, each a
duplicate this branch did not create; `server/src/edit.ts`'s document lookup,
whose `text === null` branch is deliberately spelled and would go half-dead.

The MCP transcripts are byte-identical before and after all three passes.
