+++ packages/format/README.md
@@ -88,7 +88,9 @@ The same module answers the one question about a day that is not about nodes at 
 
 A day's nodes and a zoomed page are built from the same `Situated` — a node, its mark, its rollup, what is standing in its way and its canonical ancestry — because that is one concept with several readers, not a shape two surfaces happened to agree on. A title torn out of its outline says nothing, wherever it is being drawn.
 
-`documents.ts` is the third such claim, about the `.md` files beside the outlines. A document's text is content — verbatim on disk, markdown at view time, nothing about it validated — so what lives here is only what has to be answered the same way twice: where a node's `doc` lands (`docOf`, resolved against the outline that named it, which is the rule the validator checks and the view links with), and what a relative `![](…)` may name (`pictureOf`, `isPicture`). That last one is asked by two packages that cannot import each other — the renderer that rewrites a picture into a URL, and the route that answers it — and two allowlists that drifted apart would mean either a broken image or a served file nobody meant to serve. `isAsset` is the same question asked once more and one step wider: what the media route may ANSWER at all, which is a page and the parts a page draws itself with (a stylesheet, a script, a font) as well as a picture, because a previewed `.html` is fetched by URL and draws itself. The two are deliberately not one list — markdown may still name a picture and nothing else, which is a rule about what markdown MEANS. `bodiedOf` is the third thing markdown can point at, and the third reading of the same arithmetic: where a relative `[…](…)` to another `.md` lands. A vault linking between its own files writes those by the dozen, and resolving one beside the file it was WRITTEN in is what keeps it from being resolved against whatever page it is being drawn on — which is the same reason `doc` is resolved against the outline that named it. Whether the directory actually holds the answer is nobody's business here: this package knows the arithmetic, and the page that opens the address already has a screen for a document that is not there. Three things here are not arithmetic, and each is one question four callers were asking separately. **`documentsIn` / `documentIn`** answer *which of the set's bodied files are documents* — the set keeps one list for both kinds, and the validator checking a `doc`, the planner refusing a `write_document` and both document reads each want the `.md` half of it; four `.filter`s of one rule is four places for it to drift, and the failure mode is a quiet disagreement about what a served directory contains. **`noSuchDocument`** is what a path that is not one is told, so `read_document` and `write_document` answer one typo with one sentence (the near miss is `didYouMean`, the same function an unknown node id gets); only the clause for a set with no near miss is per-verb, so only that is passed in. **`firstLine`** is the closest a `.md` has to a title — its first non-blank line with the heading marks off — a DERIVATION because a document has no record for a name to be written on; it was `@olai/web`'s while the browser was the only caller, and moved here whole when `list_documents` wanted the same answer the row beside a `doc` draws. **`bytesOf`** is beside it as the other thing sayable about a document without reading it: UTF-8 bytes, counted rather than encoded, since encoding a directory to measure it is what a listing exists not to do.
+`documents.ts` is the third such claim, about the `.md` files beside the outlines. A document's text is content — verbatim on disk, markdown at view time, nothing about it validated — so what lives here is only what has to be answered the same way twice: where a node's `doc` lands (`docOf`, resolved against the outline that named it, which is the rule the validator checks and the view links with), and what a relative `![](…)` may name (`pictureOf`, `isPicture`). That last one is asked by two packages that cannot import each other — the renderer that rewrites a picture into a URL, and the route that answers it — and two allowlists that drifted apart would mean either a broken image or a served file nobody meant to serve. `isAsset` is the same question asked once more and one step wider: what the media route may ANSWER at all, which is a page and the parts a page draws itself with (a stylesheet, a script, a font) as well as a picture, because a previewed `.html` is fetched by URL and draws itself. The two are deliberately not one list — markdown may still name a picture and nothing else, which is a rule about what markdown MEANS. `bodiedOf` is the third thing markdown can point at, and the third reading of the same arithmetic: where a relative `[…](…)` to another `.md` lands. A vault linking between its own files writes those by the dozen, and resolving one beside the file it was WRITTEN in is what keeps it from being resolved against whatever page it is being drawn on — which is the same reason `doc` is resolved against the outline that named it. Whether the directory actually holds the answer is nobody's business here: this package knows the arithmetic, and the page that opens the address already has a screen for a document that is not there. Three things here are not arithmetic. **`documentsIn` / `documentIn`** answer *which of the set's bodied files are documents* — the set keeps one list for both kinds, and the validator checking a `doc`, the planner refusing a `write_document` and both document reads each want the `.md` half of it; four `.filter`s of one rule is four places for it to drift, and the failure mode is a quiet disagreement about what a served directory contains. **`firstLine`** is the closest a `.md` has to a title — its first non-blank line with the heading marks off — a DERIVATION because a document has no record for a name to be written on; it was `@olai/web`'s while the browser was the only caller, and moved here whole when `list_documents` wanted the same answer the row beside a `doc` draws. **`bytesOf`** is beside it as the other thing sayable about a document without reading it: its size in UTF-8 bytes rather than in the UTF-16 units JavaScript holds it in, so olai reports the number every other tool a person has reports for the same file.
+
+What is deliberately NOT here is the SENTENCE a path that is not a document is refused with. This package declares what a refusal *is* (`failure.ts`); `@olai/ops` is the only layer that raises one, and it keeps that one too (`noSuchDocument`, beside `notFound`), because a floor that started composing agent-facing prose would be a second voice for the same "no". The facts the sentence is made of — `documentsIn` and `didYouMean` — are here, which is the line: facts down, voice up.
 
 ## Two readings of a set, and what they differ by
