+++ packages/ops/src/plan.ts
@@ -40,7 +40,9 @@ import {
   derive,
   type Derived,
   didYouMean,
+  type Document,
   documentIn,
+  documentsIn,
   drawingPath,
   DOCUMENT_EXT,
   isMirror,
@@ -75,7 +77,6 @@ import {
   type Capture,
   type Minted,
   NESTING,
-  noSuchDocument,
   type WriteRequest as Request,
 } from "@olai/format"
 import { Result } from "effect"
@@ -410,6 +411,47 @@ export const notANode = (id: string, target: string): OpFailure =>
     reason: `\`${id}\` is a mirror — a second placement of \`${target}\`, ` +
       `not a node of its own. Name \`${target}\` instead.`,
   })
+
+/**
+ * WHAT A PATH THAT IS NOT A DOCUMENT IS TOLD — one sentence, for every verb
+ * that can be handed one.
+ *
+ * {@link notFound}'s counterpart for the other thing an op can name. A
+ * `write_document` and a `read_document` refuse the same miss, and each built
+ * the same near-miss list out of the same set and then wrote the same sentence
+ * before this was one function: a caller who mistypes a path once should not
+ * learn two different things about it depending on which verb the typo landed
+ * at. The near miss is `didYouMean` — the same function an unknown node id
+ * gets, one moment earlier than the validator would give it.
+ *
+ * WHAT EACH CALLER KEEPS is the clause for a set with no near miss at all,
+ * because there the useful thing to say genuinely differs: a read is pointed
+ * at the listing, a write at the verb that starts a document. That is the one
+ * per-verb part, so it is the one part passed in.
+ *
+ * HERE rather than on the floor, which is where it was first written. The
+ * format declares what a refusal IS ({@link ../../format/src/failure.ts}) and
+ * this layer is the only one that raises one — a package that started
+ * composing agent-facing prose would be a second voice for the same "no".
+ * `documentsIn` and `didYouMean` are down there, and they are the two FACTS
+ * this sentence is made of.
+ *
+ * It takes the BODIED LIST rather than a scope: {@link ./query.ts} calls it
+ * over a bare set, which is what a read has.
+ */
+export const noSuchDocument = (
+  bodied: ReadonlyArray<Document>,
+  file: string,
+  instead: string,
+): OpFailure => {
+  const near = didYouMean(file, documentsIn(bodied).map((entry) => entry.file))
+  return new NotFoundFailure({
+    reason: near === ""
+      ? `\`${file}\` is not a document under the served directory — ${instead}`
+      : `\`${file}\` is not a document under the served directory${near}`,
+    named: file,
+  })
+}
 
 /** The record with this id, or the refusal that says so. A MIRROR is not an
  *  answer: it is a second placement of a node that lives elsewhere, and every
@@ -4252,7 +4294,7 @@ const planWriteDocument = (
       noSuchDocument(
         scope.set.documents,
         request.file,
-        "\`create_document\` is what starts one",
+        "`create_document` is what starts one",
       ),
     )
   }
